Move get_quick_insights into seperate file
